requirements.txt
# build
protobuf

# test
pytest
pytest-cov
tox
flake8
GitPython
hypothesis

#docs
Sphinx
sphinx_rtd_theme
sphinxcontrib.mermaid
